requirements.txt
pandas
requests
bs4
atlassian-python-api
python-dotenv